major bug fix wrong weather var types

+++ docker/mattermost-bots/src/main/java/edu/dhbw/stuttgart/tinf20b/cloud/cloudproject/weather/WeatherConnector.java
@@ -16,8 +16,8 @@ public interface WeatherConnector {
 
 
     @GetMapping("/data/2.5/onecall")
-    WeatherForecast getWeather(@RequestParam long lat,
-                               @RequestParam long lon,
+    WeatherForecast getWeather(@RequestParam double lat,
+                               @RequestParam double lon,
                                @RequestParam String appid,
                                @RequestParam String exclude,
                                @RequestParam String units);
