+++ docker/mattermost-bots/src/main/java/edu/dhbw/stuttgart/tinf20b/cloud/cloudproject/weather/model/WeatherLocation.java
@@ -12,7 +12,7 @@ import lombok.NoArgsConstructor;
 public class WeatherLocation {
 
     private String name;
-    private long lat;
-    private long lon;
+    private double lat;
+    private double lon;
 
 }
